Add cdt-gdb-adapter to bundledDependencies
For Theia, we would like to create archives that we can download,
extract and use as-is.  For it to work, the archive needs to include
cdt-gdb-adapter, which is in node_modules.  Adding it to
bundledDependencies will make "pack" bundle it in the cdt-gdb-vscode
tarball.

For some reason, "yarn pack" gives me an error:

    $ yarn pack
    yarn pack v1.12.3
    error An unexpected error occurred: "Cannot find module \".\"".

but it's still possible to use "npm pack" successfully as a workaround.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>