Adding the ability to configure serial port output to the debug console
This commit introduces a set of parameters that allow us to configure a
serial line in the adapter to print UART output from on-board/emulation
debug to the debug console. These parameters include the TCP port to
hook onto (for emulation programs outputting UART output), or the serial
port on the host-machine to capture UART output from. For the serial
line, we also configure the baud rate, stop bits, handshaking method,
etc.