Adding the ability to configure serial port output to the debug console (#91)
This commit introduces a set of parameters that allow us to configure a
serial line in the adapter to print UART output from on-board/emulation
debug to the debug console. These parameters include the TCP port to
hook onto (for emulation programs outputting UART output), or the serial
port on the host-machine to capture UART output from. For the serial
line, we also configure the baud rate, stop bits, handshaking method,
etc.

The main PR for the UART in the cdt-gdb-adapter is:

- https://github.com/eclipse-cdt-cloud/cdt-gdb-adapter/pull/271

There are additional commits following that one too, including:

- https://github.com/eclipse-cdt-cloud/cdt-gdb-adapter/pull/279
- https://github.com/eclipse-cdt-cloud/cdt-gdb-adapter/pull/280
- https://github.com/eclipse-cdt-cloud/cdt-gdb-adapter/pull/281